requirements.txt
flask==2.3.2
flask_cors=3.0.10
python-json-logger==2.0.7
celery==5.2.7
redis==4.5.5
flask_sqlalchemy==3.0.5
pymysql==1.1.0